Restructure to allow feedforward/back drive control and kalman drive control be used in Module
